update logic for GraphicSystem
